redis creators

+++ publisher/src/main/java/com/javarush/stepanov/publisher/service/entityservice/CreatorService.java
@@ -11,6 +11,9 @@ import org.springframework.data.domain.Pageable;
 
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Objects;
+
+import static java.util.stream.Collectors.toList;
 
 @Service
 @AllArgsConstructor
@@ -21,11 +24,21 @@ public class CreatorService {
     private final CreatorRedisRepo redisRepo;
 
     public List<Creator.Out> getAll() {
-        return repo
-                .findAll()
+        if (redisRepo.isAllCollectionInRedis()) {
+            return redisRepo.findAll();
+        }
+        List<Creator.Out> listResult = repo.findAll()
                 .stream()
                 .map(mapper::out)
+                .filter(Objects::nonNull)  // Отсеиваем null
+                .peek(x -> {  // peek() вместо map(), т.к. это side-эффект
+                    if (!redisRepo.exists(x.id)) {
+                        redisRepo.save(x.id, x);  // Сохраняем каждый элемент
+                    }
+                })
                 .toList();
+        redisRepo.setAllCollectionInRedis(true);
+        return listResult;
     }
 
     public List<Creator.Out> getAll(int pageNumber, int pageSize) {
@@ -51,14 +64,16 @@ public class CreatorService {
 
     public Creator.Out create(Creator.In input) {
         Creator creator = mapper.in(input);
-        List<Creator> list= repo.findAll();
-        for(Creator creatorInList : list ){
+        List<Creator.Out> list= getAll();
+        for(Creator.Out creatorInList : list ){
             if((creatorInList.getId().equals(creator.getId()))||(creatorInList.getLogin().equals(creator.getLogin()))){
                 throw new NoSuchElementException();
             }
         }
-        return mapper.out(
+        Creator.Out result = mapper.out(
                 repo.save(creator));
+        redisRepo.save(result.getId(), result);
+        return result;
     }
 
     public Creator.Out update(Creator.In input) {
@@ -69,10 +84,16 @@ public class CreatorService {
          existing.setPassword(input.getPassword());
          existing.setFirstname(input.getFirstname());
          existing.setLastname(input.getLastname());
-        return mapper.out(repo.save(updated));
+        Creator.Out result = mapper.out(repo.save(updated));
+        redisRepo.save(result.getId(), result);
+        return result;
     }
 
     public void delete(Long id) {
+        if (repo.findById(id).isEmpty()) {
+            throw new NoSuchElementException("Creator not found with id: " + id);
+        }
+        redisRepo.delete(id);
         repo.deleteById(id);
     }
